refactor: move stacks to constants
